Fixed issues found when playing cents game (#645)
* Working on cents game

* First round of work to conver to cents game

* Fixed issues found when converting to cents game

* Updated build number

* Reverted game server version number

* Fixed few more issues with cents game

* fixes the key error in parsing ResultWinner

* minor fix to the showWinner function - thus showing results for all winners

* fixes a minor bug

* fixes hi-lo hand issues

* fixes background color for replay hand dialog

* minor change

* clipping the backdrop from the replay hand dialog widget

* Fixed issues found in cents game

+++ lib/models/game_play_models/provider_models/game_state.dart
@@ -927,7 +927,7 @@ class GameState {
         name: playerInSeatModel.name,
         seatNo: playerInSeatModel.seatNo,
         playerUuid: playerInSeatModel.playerId,
-        stack: playerInSeatModel.stack.toInt(),
+        stack: playerInSeatModel.stack,
       ));
     });
     if (notify) {
@@ -1152,10 +1152,10 @@ class GameState {
   }
 
   void updatePlayersStack(Map<dynamic, dynamic> stackData) {
-    Map<int, int> stacks = Map<int, int>();
+    Map<int, double> stacks = Map<int, double>();
 
     for (final key in stackData.keys) {
-      stacks[key.toInt()] = stackData[key].toInt();
+      stacks[key.toInt()] = stackData[key];
     }
     // stacks contains, <seatNo, stack> mapping
     stacks.forEach((seatNo, stack) {
